requirements.txt
behave==1.2.6
selenium==3.141.0
assertpy==1.1
xmlrunner==1.7.7
PyHamcrest==1.9.0
